Fluently-Playground/requirements.txt
torch
diffusers
transformers
gradio==4.19.2
pillow
omegaconf
peft
numpy